Add files via upload
This was a simple website project idea for an online shopping service, which would provide cheaper and reliable shipping options, better prices for clothes and also help those in need by donating some of the batches produced to those in need. It's not yet where I want it to be, that's why I'm welcoming ideas from different designers and developers to help out on it.